requirements.txt
flask>=3.0.0
gunicorn>=22.0.0
python-dotenv>=1.0.0
requests>=2.31.0
anthropic>=0.40.0
APScheduler>=3.10.0
markdown>=3.5.0
